% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrepancy.r
\name{discrepancy}
\alias{discrepancy}
\title{calculate the discrepancy for a single sample}
\usage{
discrepancy(
  sample_data = NULL,
  imputed_data = NULL,
  procedure = "interpolation",
  model = FALSE,
  ct_type = mean
)
}
\arguments{
\item{sample_data}{a data.table object of the original sample data}

\item{imputed_data}{a list of data.table objects with the imputed data}

\item{procedure}{the procedure used}

\item{model}{is it the discrepancy for a model or not
load dependencies}
}
\description{
calculate the discrepancy for a single sample
}
