% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_data.R
\name{.complete_data}
\alias{.complete_data}
\title{cleaning miceRanger}
\usage{
.complete_data(data_frame)
}
\arguments{
\item{data_frame}{A data.table/data.frame object
containing the complete imputed data from miceRanger imputation}
}
\value{
df_miceranger list of data.frames
}
\description{
This is a function that takes a list of miceRanger data.table outputs
and uses miceRanger::completeData on each list element to return a
data.frame of the completed data after miceRanger imputation. After
doing this, it then takes the list of completeData data.frames and
collapses them into one data.frame with a dataset id viarable.
}
\details{
This is a private function used to clean miceRanger output from
the impute function.
}
\examples{
df_miceranger <- miceranger(df_missing)
df <- complete_data(
  data_frame = df_miceranger
)
}
\seealso{
\code{\link[=impute]{impute()}}

Other impute: 
\code{\link{impute}()}
}
\concept{impute}
