% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{impute}
\alias{impute}
\title{impute the data}
\usage{
impute(data_frame, m = 10, package = NULL, meth = NULL)
}
\arguments{
\item{data_frame}{The data.frame with missingness}

\item{m}{The number of imputations to perform}

\item{package}{The package to use for the imputation}

\item{meth}{The method of imputation}
}
\value{
list of data.tables
}
\description{
This function takes incomplete data and will impute it using
either the mice, Amelia, or miceRanger packages. Once it has
finished imputing the data, it will then return a
data.table/data.frame object with the complete data from the
imputation and will record which rows were from which
imputation iteration.
}
\details{
This function relies on the private complete_data function
to clean the miceRanger data.
}
\examples{
list_imputed <- impute(
  data_frame = df_missingness
  , m = 10
  , package = "mice"
  , meth = "mean"
)
length(list_imputed) # returns 10. 1 for each imputed data.table.
}
\seealso{
\code{\link[=.complete_data]{.complete_data()}}

Other impute: 
\code{\link{.complete_data}()}
}
\concept{impute}
