% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate}
\alias{simulate}
\title{simulate}
\usage{
simulate(N = 1e+06)
}
\arguments{
\item{N}{the size of the population data.frame}
}
\value{
df_population.
A data.table object containing the population data.
}
\description{
This is a function that simulates five columns and returns a
data.table object. The equations for the simulation of the
columns are the following.
\itemize{
\item Variable A: gamma(shape=1)
\item Variable B: binomial(0.6)
\item Variable X: 0.2 * A + 0.5 * B + normal(0, 10)
\item Variable Z: 0.9 * A * B + normal(0, 10)
\item Variable Y: 0.6 * Z * Z + normal(0, 10)
}
}
\examples{

df_population <- simulate(N = 100)
nrow(df_population) # returns 100

df_population <- simulate(N = 1000)
nrow(df_population) # returns 1000
}
